Consolidate fit classes to one module (#95)
