+++ gpfit/tests/t_LM.py
@@ -2,7 +2,7 @@
 import unittest
 from numpy import arange, newaxis
 from gpfit.levenberg_marquardt import levenberg_marquardt
-from gpfit.max_affine import max_affine
+from gpfit.classes import max_affine
 
 
 def rfun(params):
